Move the demo function into the App
- handles interactive run by `java -cp app/build/classes/java/main dev.madpang.App`
- handles gradle invoking by `./gradlew run  --args=<your-test-file-path>`
